Mova 104 - Implement the DetailScreen #9
- Implemented detail screen.

+++ shared/src/commonMain/kotlin/feature_detail/data/remote/tv/dto/SeasonDto.kt
@@ -12,5 +12,5 @@ data class SeasonDto(
     val overview: String?,
     @SerialName("poster_path") val posterPath: String?,
     @SerialName("season_number") val seasonNumber: Int?,
-    @SerialName("vote_average") val voteAverage: Int?
+    @SerialName("vote_average") val voteAverage: Double?
 )
